refactor dn() to a class, match filename and fix in main
